+++ CHANGELOG.md
@@ -1,6 +1,16 @@
 # Change Log
 All notable changes to this project will be documented in this file.
 
+# Version 1.1.0
+
+#### Added
+* Added setUserConsent method for GDPR
+
+# Version 1.0.1
+
+#### Added
+* Fix banner fail to load event
+
 # Version 1.0.0
 
 #### Added